requirements.txt
django
gunicorn
django-decouple
django-widgets-improved
requests
pillow
whitenoise
django-taggit